+++ src/facade/add_handler.rs
@@ -1,9 +1,11 @@
-use clap::ArgMatches;
 use crate::dao::read_json::{fetch_todos, parse_json_from_string};
 use crate::dao::write_json::serialize_model_to_json;
 use crate::util::display::display_todo_added_msg;
 use crate::util::enums::TodoStatusType;
-use crate::util::helpers::{get_current_date_time_iso, read_file_from_path};
+use crate::util::helpers::{
+    get_current_date_time_iso, get_tag_annotation_from_string, read_file_from_path,
+};
+use clap::ArgMatches;
 // custom
 use crate::util::models::{FileRequestResponse, Todo};
 
@@ -13,14 +15,15 @@ use crate::util::models::{FileRequestResponse, Todo};
  * > jd add
  */
 
-fn create_todo(text: String) -> Todo {
+fn create_todo(text: String, tag: String) -> Todo {
     let todos: Vec<Todo> = fetch_todos();
     let count = todos.len();
     let record: Todo = Todo {
         id: (count + 1) as i64,
         desc: text.to_string(),
+        tag: tag.to_string(),
         status: TodoStatusType::Pending.to_string(),
-        modified: get_current_date_time_iso()
+        modified: get_current_date_time_iso(),
     };
     record
 }
@@ -29,14 +32,13 @@ fn create_todo(text: String) -> Todo {
  * "add "some todo item"" subcommand handler
  * @param {&ArgMatches} args
  */
-pub fn handle_add_todo (args: &ArgMatches) {
+pub fn handle_add_todo(args: &ArgMatches) {
     let argument = args.get_one::<String>("todo").unwrap();
-    let record: Todo = create_todo(argument.to_string());
+    let (_, tag, _) = get_tag_annotation_from_string(&argument);
+    let record: Todo = create_todo(argument.to_string(), tag);
     let json_string: String = read_file_from_path();
     let mut model: FileRequestResponse = parse_json_from_string(json_string);
     model.update_todo_list(record);
     serialize_model_to_json(model);
     println!("{}", display_todo_added_msg());
 }
-
-