+++ src/util/models.rs
@@ -6,7 +6,6 @@ use serde_derive::Serialize;
 pub struct FileRequestResponse {
     pub tags: Vec<String>,
     pub todos: Vec<Todo>,
-    pub reminders: Vec<Reminder>,
     pub tasks: Vec<Task>,
 }
 
@@ -15,7 +14,6 @@ impl FileRequestResponse {
         Self {
             tags: vec![],
             todos: vec![],
-            reminders: vec![],
             tasks: vec![],
         }
     }
@@ -30,19 +28,8 @@ impl FileRequestResponse {
 pub struct Todo {
     pub id: i64,
     pub desc: String,
+    pub tag: String,
     pub status: String,
-    pub modified: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Reminder {
-    pub id: i64,
-    pub desc: String,
-    pub status: String,
-    pub due: String,
-    #[serde(rename = "reminder_type")]
-    pub reminder_type: String,
     pub modified: String,
 }